Implemented new logging system focusing on message flow
Diagnostic messages related to message-passing now use a table-based
--- and color-coded --- format that allows for quickly gaining an
understanding of what ProxyRMI is doing behind the scenes.

While there are (of course!) still improvements that could be made, this
is a great step in the right direction.  As usual, verbose
diagnostics/logging is enabled via the `#verbose=` instance methods.

Warning: uses ANSI terminal escape sequences.